install.sh: support for symlinking directory structures
Adds support for specifying an origin and a destination directory
in the dotfiles array. When specified, all files under the origin
directory will be symlinked into the same directory structure
under the destination directory. Directories themselves are not
symlinked, only the files within. Non-existing directories are
created into the destination directory.