install.sh: add support for running scripts
Scripts can be placed in a script directory, after which they
can be executed through the --script option.

Scripts are sourced so all variables and functions available
in install.sh are available also for the scripts.

This change is to be able to generalize install.sh so that installation
of specific tools etc can be put in to scripts instead of being
part of install.sh